+++ ncam-log.c
@@ -697,7 +697,7 @@ void cs_statistics(struct s_client *client)
 		/* statistics entry start with 's' to filter it out on other end of pipe
 		 * so we can use the same Pipe as Log
 		 */
-		snprintf(buf, sizeof(buf), "s%02d.%02d.%02d %02d:%02d:%02d %3.1f %s %s %d %d %d %d %d %d %d %ld %ld %02d:%02d:%02d %s %04X@%06X:%04X %s\n",
+		snprintf(buf, sizeof(buf), "s%02d.%02d.%02d %02d:%02d:%02d %3.1f %s %s %d %d %d %d %d %d %d %lld %lld %02d:%02d:%02d %s %04X@%06X:%04X %s\n",
 				 lt.tm_mday, lt.tm_mon + 1, lt.tm_year % 100,
 				 lt.tm_hour, lt.tm_min, lt.tm_sec, cwps,
 				 client->account->usr,
@@ -709,8 +709,8 @@ void cs_statistics(struct s_client *client)
 				 client->cwignored,
 				 client->cwtout,
 				 client->cwtun,
-				 client->login,
-				 client->last,
+				 (long long)client->login,
+				 (long long)client->last,
 				 fullhours, mins, secs,
 				 get_module(client)->desc,
 				 client->last_caid,
